input: refactor ScmOverrides
Moving scmOverrides class into a separate module next to the real scms.
