+++ pym/bob/input.py
@@ -16,7 +16,7 @@
 
 from . import BOB_VERSION, BOB_INPUT_HASH, DEBUG
 from .errors import ParseError
-from .scm import CvsScm, GitScm, SvnScm, UrlScm
+from .scm import CvsScm, GitScm, SvnScm, UrlScm, ScmOverride
 from .state import BobState
 from .tty import colorize, WarnOnce
 from .utils import asHexStr, joinScripts, sliceString, compareVersion, binLstat
@@ -468,42 +468,6 @@ class PluginState:
         :param bob.input.Package packages: The created package
         """
         pass
-
-class ScmOverride:
-    def __init__(self, override):
-        self.__match = override.get("match", {})
-        self.__del = override.get("del", [])
-        self.__set = override.get("set", {})
-        self.__replaceRaw = override.get("replace", {})
-        self.__init()
-
-    def __init(self):
-        self.__replace = { key : (re.compile(subst["pattern"]), subst["replacement"])
-            for (key, subst) in self.__replaceRaw.items() }
-
-    def __getstate__(self):
-        return (self.__match, self.__del, self.__set, self.__replaceRaw)
-
-    def __setstate__(self, s):
-        (self.__match, self.__del, self.__set, self.__replaceRaw) = s
-        self.__init()
-
-    def __doesMatch(self, scm):
-        for (key, value) in self.__match.items():
-            if key not in scm: return False
-            if not fnmatch.fnmatchcase(scm[key], value): return False
-        return True
-
-    def mangle(self, scm):
-        if self.__doesMatch(scm):
-            scm = scm.copy()
-            for d in self.__del:
-                if d in scm: del scm[d]
-            scm.update(self.__set)
-            for (key, (pat, repl)) in self.__replace.items():
-                if key in scm:
-                    scm[key] = re.sub(pat, repl, scm[key])
-        return scm
 
 def Scm(spec, env, overrides):
     # resolve with environment
