+++ pym/bob/scm/scm.py
@@ -1,0 +1,54 @@
+# Bob build tool
+# Copyright (C) 2016  Jan Klötzke
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import fnmatch
+import re
+
+class ScmOverride:
+    def __init__(self, override):
+        self.__match = override.get("match", {})
+        self.__del = override.get("del", [])
+        self.__set = override.get("set", {})
+        self.__replaceRaw = override.get("replace", {})
+        self.__init()
+
+    def __init(self):
+        self.__replace = { key : (re.compile(subst["pattern"]), subst["replacement"])
+            for (key, subst) in self.__replaceRaw.items() }
+
+    def __getstate__(self):
+        return (self.__match, self.__del, self.__set, self.__replaceRaw)
+
+    def __setstate__(self, s):
+        (self.__match, self.__del, self.__set, self.__replaceRaw) = s
+        self.__init()
+
+    def __doesMatch(self, scm):
+        for (key, value) in self.__match.items():
+            if key not in scm: return False
+            if not fnmatch.fnmatchcase(scm[key], value): return False
+        return True
+
+    def mangle(self, scm):
+        if self.__doesMatch(scm):
+            scm = scm.copy()
+            for d in self.__del:
+                if d in scm: del scm[d]
+            scm.update(self.__set)
+            for (key, (pat, repl)) in self.__replace.items():
+                if key in scm:
+                    scm[key] = re.sub(pat, repl, scm[key])
+        return scm
